cursor_git/djangoProject/requirements/dev.txt
-r base.txt

pytest-django==4.1.0
mock==4.0.3
factory-boy==3.2.0
django-debug-toolbar
faker-vehicle==0.2.0
Faker==8.11.0
psycopg2-binary

